% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airway_small}
\alias{airway_small}
\title{A small example dataset for SummarizedExperiment}
\format{
SummarizedExperiment with 4001 rows
}
\source{
the airway data
}
\usage{
airway_small
}
\description{
A small example dataset for SummarizedExperiment
}
\keyword{datasets}
